Set the default frequency to 300
Linux Perf tool's default frequency is 4000. However, this number is not
a problem for them because it's based on 3rd party process monitoring.

In some systems, if hpcrun sets 4000 samples per second, the kernel may
adjust the period into ridiculously low number such as 20 (for cycle
event, it means the overflow is for every 20 cycles !).

There is no parameter to force Linux kernel not to adjust the period
lower than a certain number. So far, using 300 samples per second is
good enough, and close to the result of perf tool (the percentage
number).