Add hpcrun flag to set the default period threshold
Instead of using the hard-coded default period (frequency 300 samples
per sec), users can define the default period in the command line:

hpcrun -c <period>  -e ....

The value of <period> is exactly the same as the value of -e option.
For instance:

-c 4000 
the default is overflow threshold 4000

-c f4000
the default threshold is 4000 samples per second

Note: -c option is similar with linux perf -c flag.