Add the full solution for unknown proc to makeStructure() and work around a couple SymtabAPI issues.  This allows getting real proc names for PLT stub functions.
SymtabAPI doesn't recognize PLT functions and mistakenly locates them
in the _init() function.  If there is no valid Symtab function for a
ParseAPI function, then create a fake one that extends to the next
function.  This requires changing the GroupMap index to VMA.

Also, if a Symtab function extends past the end of its region, then
truncate it to the end of the region.