Add a dummy kernel blocking module for old linux kernel
During the configuration we detect if the kernel has attributes to build kernel blocking (which is provided by kernel 4.3 or newer).
If the test is true, we build using the true kernel blocking module.
If false, we use the empty module (kernel_blocking_stub.c).

This simplifies the code as we avoid conditional preprocessor.