hpcprof-mpi: add finalize phase for perf event statistics
One perf event statistics is the mean of period sampling (for
frequency-based sampling, the period varies. This info will help users
to know the average of the sampling threshold).

With hpcprof-mpi, we need to compute the average of "the mean of period
sampling" across multiple profiles. The simplest way to do this is to
compute the average after the merge on the root rank.