minor update: rename perf event attributes in xml
shorten the name of attributes removing the first letter, e.g. instead of 'pep', renamed to 'ep'.
The advantage of this is to make the first letter of attributes in metric to be unique. The letter 'p' is already taken by attribute 'partner'.

other minor update: clean-up code in perfmon-util