Fix problem of profiling with many events with high rate
The problem in this situation is that once a thread get sampled from a counter, other counters still continue to run. This may cause another signal in the middle of signal handler.
Although we already have mechanism to deal with it (with hpcrun_safe_enter_async), it's better to freeze all counter while we are inside a signal handler. We will re-enable the counters once we finish processing.

So far it's been tested with around 20ish counters, and the number of blocked samples is around 13 (out of 20,000 samples).