First draft datacentric: heap allocation
major changes:
- The datacentric is based on perf event. Using event-custom module, the
datacentric is called and managed by perf events sample-source, instead
of creating its own sample-source. (it's a bit like BLOCKTIME event).
The reason for this is we'll use perf event to gather information on ip
address and other perf event information. It's more logical to be
related to perf events than being an independent sample-source.

- applying some Xu's code for patching the cct. Changing/updating cct is
very tricky in hpcrun. I am not sure the proper way to do this. This
requires some surgeries in hpcrun cct mechanism.