Generalize customizing cct in hpcrun
We need a framework that allows to modify cct in hpcrun. For instance,
with linux perf sample sources, we need to add kernel call path. For
datacentric, we need to add tree root and path to the allocation. For
java, we need to replace the cct with jvm's callpath.

Instead of hacking the backtrace module (cct_insert_backtrace.*), we
should indicate when calling hpcrun_sample_callpath() if we want to
to call our own custom cct before or after the unwinding (or both !).