Code reorganization
A datacentric module consists of 3 parts:

1. memory address (retrieved from malloc wrap or data-overrides.* and
hpcfnbounds static memory)
2. memory database (data_tree.*)
3. precise events like ibs, pebs and marked events. This part is very
hardware specific.

This commit attemps to modularize the 3rd part so that we can adopt
datacentric approach based on the supported architecture.
For instance, if during the configuration we determine that the hardware
supports ibs-op, the datacentric sample-source will use "plugin" from
ibsop.c.