Datacentric: Working version of simple malloc wrapper
This is a working version of a simple *alloc wrapper without using header/footer embedded information.
The performance may not be good as it's always to splay tree whenever we free a memory (compared to in-band/embedded information which has the information already in hand).

Currently it stores the allocation information in a local thread's splay tree to avoid lock serialization. However, if another thread attempts to free another thread's allocated memory, it requires to poke each thread, which causes performance degradation.

However, this version is working fine when profiling with hpcstruct, while using in-band information causes infinite loop (see issue #24 on https://github.com/HPCToolkit/hpctoolkit/issues/24 )