"working" version of address-centric using page-faults
Overhaul the event_custom module so that it can accommodate both kernel blocking time plugin and datacentric plugin.

Instead of using mprotect(), we use perf_event's page-faults to retrieve the ip-address and the effective address. This commit use the effective address and compare with the existing memory allocation from malloc.
Currently it always incrementally add the address instead of setting the first-touch address.

Note: The IP provided by perf events, is not used yet.