test pmu for profilability in attribute leevel
when users run "hpcun -L", we only test on the pmu level. However, some pmus require to include an attribute to be profilale.
For this case, we should test in the attribute level (if possible).
If a counter is not profilable, we still keep showing in the list, but need to mark that the counter may not be profilable.