Working version of address-centric profiling
There are 2 components in address-centric profiling:

- data-overrides.*: overrides malloc and free to capture the address and
the size of the allocated memory

- datacentric.*   : (file name has to be renamed) Using PAGE-FAULT perf
events to capture the first touch of an allocated memory. Once an event
is delivered, we see if the address is in the database. If this is the
case, we store the address of the first touch.

Minor update:
- simplify the tsmg output. Originally it's too many noisy message.