hpcfnbounds: add query to static variables
Needed for datacentric: query information on static variables in a load module.

the server:
- Adding -m flag in hpcfnbounds to distinguish with the default query: function bounds
- has a specific do_query function for searching static variables.
- using dyninst's SYMBOL::ST_OBJECT to get the addresses.
- add new object: variable-entries which is similar to function-entries

the client (hpcrun):
- when there's a new load module, immediately send a query to ask for the static variables
- if the load module has the info, store it in a database (splay tree).

Notes:
At the moment, we'll collect all the addresses of the variables and store them into a splay tree, together with dynamic variables.
To know if a variable is a dynamic (allocated) or static, we need to check if the datainfo_s has the context field. In a static variable, this field is basically empty (we need to define perhaps with other thing).