Add the libdw case to dwarf_eh_frame_info() in fnbounds.  Run the libdwarf functions via dlopen() and dlsym() as a way to keep them isolated from libdw and dyninst.
Currently, we re-dlopen(libdwarf.so) for each file and close it before
invoking the callback function (to better separate libdwarf and
libdw).  But we may be able to relax that and hold an open handle
across files.

At this point, the libdw case runs with fnbounds and hpcrun.  But the
build is not quite 100% clean over the compression prereqs for
elfutils and libdw.