hpcprof: Working draft merging lm <vmlinux.xxxxx> into a single lm
To merge multiple vmlinux.xxxx load module, there are 2 steps:

1. In CallPath-Profile.cpp, before writing the xml header, we check if
there are multiple vmlinux. If this is the case, the first one will be
the "leader" and substitute other vmlinux.

2. In CCT-Tree.cpp, before writing a cct PF, we need to check if the
load module is to be subtituted or not. If it is the case, we substitute
with the id of the "leader"

TODO:

- to interact between CallPath-Profile.cpp and CCT-Tree.cpp, we use a
global variable. This may cause a problem in the future. Using
stdatomic.h somehow doesn't compile. But using std's atomic will
conflict with our stdatomic. Not sure how to fix this.

- need to clean-up the code. Should merge with vdso branch.