Bug fix issue #45 (strange kernel chains)
The bug occurs when profiling with a perf_event event and traces.
Reason: we always assume there is no cct modification when calling hpcrun_backtrace2cct() and assume the return leaf_func is a direct descendant of sample node. (the trace will use leaf_func as their frame).
However, this is not true as deep inside hpcrun_backtrace2cct(), we add kernel call chains. Hence the variable leaf_func is not anymore direct descendant of the sample node.

Fix: Using sample node as the trace frame. This works with fib and amg2006 (apparently). Need more tests to validate.

See bug description at:
https://github.com/HPCToolkit/hpctoolkit/issues/45