Add unit test and fixing a bug
Add an embedded unit testa. It needs to build manually at the moment by adding -DHPCTOOLKIT_UNIT_TEST in the compiler flag. In the future, we need to think a proper way to run unit tests.

Fixing bug when decompressing a compressed kallsyms file: need to rewind the file stream pointer at the beginning before reading it.