Use lzma compression by default instead of zlib
HPCToolkit externals provides both lzma and zlib. Since lzma is already used by some parts in hpctoolkit (libelf I think), it is preferable to use lzma instead of zlib.

We need the compression/decompression in two tools: hpcrun (if perf_events is enabled) and hpcprof/hpcprof-mpi. However, due to complex dependencies, other tools like hpcstruct, hpclump and hpcproftt have to add liblzma as well.
The reason is that somewhere deep in hpcstruct, it requires BinUtils::LM class which then requires KernelSymbols class, which the include compression library. And since the compression is in support-lean package, other tools like hpclump also requires the compression library.

In the future, we need to clean-up the code so that hpcstruct doesn't have to include compression library.