Temporary fix for libhpcrun.o and liblzma.a
Somehow, Makefile in hpcrun adds prof-lean and support-lean libraries after the definition of libhpcrun_o_* :

 ... liblzma.a --whole-archive ../../../src/lib/prof-lean/.libs/libHPCprof-lean.a ../../../src/lib/support-lean/.l
ibs/libHPCsupport-lean.a --no-whole-archive

This causes a routine needed in support-lean that is defined in liblzma not included in libhpcrun.o. Hence, this causes link failure with hpclink (since there's undefined routine).
To fix this temporarily, we force to add liblzma.a (installed in externals package installation directory) into hpclink at the end of the linkage process.

This is not a proper solution, but at the moment it fixes the problem.