Merge some codes from VDSO branch
VDSO branch has a better infrastructure to handle custom function bounds.
This commit although grab a lot of codes from the branch, but only pick related codes to SimpleSymbols to handle kernel and vdso symbols.

The code for procedure redundancy elimination also needs to be adapted to handle both vdso and vmlinux. However, codes for vdso are still not merged.