fix corner case for itimer sample source.
handle the case when a thread receives a shootdown signal while
handling a timer interrupt. after a thread shootdown, the timer
will be turned off and freed. it cannot restart. so, if
restarting the timer fails, it is OK if the timer is marked
as unitialized and no longer available because of a shootdown.

bug comments:
a thread waiting at an OpenMP barrier was handling a sample
while receiving a shootdown signal, had its timer
turned off + deallocated, failed to restart the timer, and
then aborted with an error message that it was unable to
start the timer.