Remove setenv in pfmu_init
There is no need to set LIBPFM_ENCODE_INACTIVE variable. This variable
is required only if we want to profile with undetected-but-supported
event.
Since hpcrun requires users to specify "detected" events, we can safely
remove this variable.

Note:
setenv is problematic.
The execution of setenv in the beginning of the program will remove
HPCRUN_EVENT_LIST. Even if we need setenv, we should execute it after we
read HPCRUN_EVENT_LIST.