Use IOC_ENABLE instead of IOC_REFRESH
Both ENABLE and REFRESH can enable a perf_event, and both are called with ioctl.
The difference is that REFRESH can enable once, and then disable automatically.
While ENABLE will make the perf_event be always enabled all the time until we explicitly disable it.

REFRESH has the advantage to be convenient. But sometimes in frequency-based sampling the sample is delivered immediately inside ioctl. When the signal handler finds that the signal is within hpcrun, it will reject immediately, thus has no time to re-enable again.

ENABLE has no problem with this, but it will deliver continuously samples until we disable it.