Fix problem in hpcprof when vmlinux is not available
There's a change in hpcrun not to compress vmlinux if the kptr_restric is not 0 (there is no need to copy or compress vmlinux where the address is all zero).
hpcprof needs to check if vmlinux is available or not. If not, we shouldn't print errors, just warning should be ok.