emit warnings rather than errors in hpcprof if kernel symbols are not available
hpcrun does not record a compressed copy of kallsyms if kptr_restrict is not 0
because all kernel symbols would be mapped to address 0, which is useless.

as a result, hpcprof needs to check if kernel symbols have been recorded or not.
if not, we shouldn't print errors; warnings are enough.