Correct the bounds checks in uw_recipe_map_repoison to ensure that we don't try to fuse a repoisoned interval with an interval bounded by the min or max sentinel values.
Change the invocation of uw_recipe_map_cmp_del_bulk_unsynch.  Formerly, we
passed the starting address of a structure that contained the start and
end of a range as if that address was the start, and also the end, of the
range.  Instead, pass the interval to uw_recipe_map_cmp_del_bulk)_unsynch,
and let it extract the start and end of the range to pass on to the skiplist
manipulation function.