adjust tracing of user-level call-path samples
avoid duplicate line mappings that occur when logging arbitrary
instruction pointers in a call path trace for a thread.
to accomplish this, when logging a call path in a trace, map its
leaf IP to the first IP in the enclosing procedure. see
comments in sample_event.c for the detailed explanation.