use binary analysis for unwinding even if errors
recent changes enforced a zero-tolerance for errors when employing
binary analysis for call stack unwinding. if we are resorting to
binary analysis and there are errors, we have no better alternative
so try to use the results.

this change is being committed after studying why binary analysis
of AMD math libraries (written in assembly code) yielded errors,
which caused us to discard its results. it appears that
the errors arise from apparently trying to apply the binary
analysis to data in the text segment that falls between routines
in the text segment. in this case, using the binary analysis
results instead of discarding them in the presence of errors
reduced unsuccessful unwinds by 80%.