Simplify perf sample record
Instead of asking many sample types that are not used (at the moment) like sampling-address, cpu and thread ID (which are to-be-used for data-centric), it's better to configure sample types as simple as possible.
The rationale of this is that some events do not support enhanced sample types (like sampling-address) and in some platforms it will fail. It is much safer to configure as minimum as possible, and let data-centric plugin to configure perf_events with much enhanced features.