Rework the configure options for Blue Gene.
Autodetect Blue Gene/Q and build for the back end by default.  Add the
single option --enable-bgq to specify building for the back end.  This
replaces three older options for Blue Gene.

The main point of this option is to turn off REALTIME and Perf Events
on the back end.  The POSIX timers don't run on the back end, and Perf
Events generates an illegal instruction.  But they pass the configure
tests on the front end.

Use the option --disable-bgq to build for the front end of a Blue Gene
machine.

Kill the last of Blue Gene/P and UPC support, which probably doesn't
build or run anymore.