hpcrun: fix precise_ip detection
Check the precise_ip attribute for each event instead of "once for all".
However, linux kernel (tested with 4.13) has an issue with this. If we check an event with a specified sampling period, linux kernel will report it's okay to sample with zero skid.
Note that if we DON'T specify the sampling period, linux kernel seems more careful/conservative.

If we look at the linux perf tool, if we specify an event, by default the skid is arbitry (no restriction). But if we don't specify an event, it will try to detect the precise_ip (like us), without specifying the sampling period.
It's unknown if this is a bug or a feature. At the moment we just follow perf tool approach.

In immediate future we should allow users to force to create an event with a more restricted skid (datacentric will require skid zero for some events).

Minor update:
- make the mmaped buffer for perf_event a little bit larger (2 pages instead of 1 page) to avoid issues when we have lot of information such as long kernel call path, ...

This commit is related to bug #70: https://github.com/HPCToolkit/hpctoolkit/issues/70