hpcrun: adjust detecting pmu in datacentric
We avoid to detect the type of processor. Instead, we guess the cpu type by attempting if an event, which probably unique to cpu type, exists or not.

TODO: datacentric event may consists of one or more pmus. We need to redesign the event_desc type so that it includes more than one events