adjust PCs used for OMPT placeholders
a kernel sample may be attached below a placeholder, causing it to be
treated as an interior node. when hpcprof processes an interior node in
a CCT, it adjusts its address by -1 to map from a return address observed
on the stack back to the preceding call instruction. make the PC for a
placeholder be 1 byte inside the function so hpcprof's interior node PC
adjustment doesn't move outside the placeholder routine.