hpcrun: fix merging cct nodes
To merge two cct nodes, we need to know if the nodes are leaf nodes or
not. However, the way we check a leaf node is not completely correct.
It's correct if the cct isn't touched or modified. But when we insert a
kernel call chain, a leaf node is not lead anymore.

We should check by verifying if a node has children or not. I think it's
more reliable that way.