datacentric: add cct node type mem_access
if a cct node is not leaf but has metrics, hpcprof always assumed that the node is call site.
this violates data centric nodes where a node that has a memory access has a child that points to the allocation node.

to solve this issue, we add a new node type "NODE_TYPE_MEMACCESS" to mark that this node is not a leaf, has metrics, and has a child node that points to the allocation node used by data-centric