draft support for short-lived threads
Add support for handling short lived threads. When a program ini a
thread, we look for available thread_data in the queue. If no data is
available, we create a new one.

On thread exit, we store the current thread data into the queue, in case
new threads need it.

to fix: only main thread will write all the profie data.