hpcrun: draft for handling metrics in short-lived threads
The way hpcrun write metrics to the file is very "local" minded. In hpcrun_get_metric_set(), we always try to find the current cct2metrics_t map with THREAD_LOCAL_MAP().
This doesn't work well when we want to retrieve the metrics from another thread. To fix this, we need to add an argument in hpcrun_get_metric_set() a pointer to cct2metrics map.
If a caller doesn't want to provide a map, it can use NULL argument (if the value is NULL, hpcrun_get_metric_set will use THREAD_LOCAL_MAP).

So far this works well with AMG kernel with 8 threads.