hpcrun: fix trace id assignment in sample event
In hpcrun, we mark a node as "retaining node" by adding 1 to the
persistent_id (by default persistent_id is always even number).

The problem is we retrieve the persistent_id BEFORE we mark the node as
the "retaining node". Hence, when we write to trace buffer, if the node
is never been marked as a "retaining node", we incorrectly write the
callpath id (which is an even number instead of odd number).

The fix is to move retrieving the persistent id AFTER we mark the node
as a "retaining node".

Minor fix: move epoch_init to inside hpcrun_thread_data_init so that we
don't need to init the epoch if we reuse thread data.