Create threads to distrobute the finalization work
At the end of the main thread, it's critical to distribute the finalization work (which consists of writing profile data, and closing hpctrace file).
To let the main thread to process this work will cause bottleneck at the end of the program.

The number of threads is the minimum of:

- the number of cores (based on get_nprocs)
- the number of thread data in the queue
- the number of "logical" threads