First pass at removing races and crashes and making the output of hpcstruct deterministic.
1. Add __thread to the static buffers in pathfind.cpp and realpath.c.
This should prevent struct from crashing with threads.

2. Sort the lists of blocks, edges and loops from parseapi.  This
should make both the struct file and debug output deterministic.
(The debug output needs to be serial, or else the interleavings are
jibberish.)

There are still nondeterministic results from parseapi, so it's hard
to tell if this fixes all the races in hpcstruct.