In hpcstruct, print separate <C> tags for call statements.
Extend StmtInfo to note which instructions are call stmts, along with
their target address, if known.  Note: a StmtInfo for a call instn is
always a single instruction, it is never merged with another StmtInfo.

In doStmtList() in Struct-Output.cpp, split the StmtInfo's into call
and non-call stmts.  The non-call <S> tags are merged by line number,
while the call <C> tags are always a single instruction.

Extend hpc-structure.dtd to add a 't' field to the <C> tag for the
target address of the call.

Note: there are small, 1-line knobs in Struct-Output.cpp to enable
printing the <C> tags and the 't' field.  These are turned off for now
because the <C> tags badly break hpcprof.