Rephrase error message when fail to initialize a perf event
In some platforms (like NOTS' head node), perf always fail to initialize hardware events, and only allow software events.
hpcrun should be able to display error message in stderr that there's a failure in an event initialization.
We cannot depend on strerror to get the error message since it's more confusing than helping in this step.