Draft using metrics to store variable address
Instead of embedding the information of variable address and the node of
the allocation/declaration, it's cleaner to store as a metric.
This case we avoid too much hassle in hpcprof-mpi. The later requires
exactly the same procedure in reading and writing cct nodes (no cct
manipulation) to be used for both reading hpcrun and pack/unpack for mpi
broadcast.

A major disadvantage of storing as a metric is that it generates
unnecessary metrics for metric-db and xml. Unless hpcrun can communicate
with hpcprof/-mpi that some metrics are only temporary.