hpcprof: print metric value based on its type
Originally, we assume all metrics have "double" type.
In datacentric, this is not the case as we want to store the address of
the variable (beginner or end or the range).

Printing an address type variable into a double value is troublesome and
is misleading for the viewer. The clean way is to print metric of
address type to an integer (or hex). For this, we need to overhaul many
parts of lib/prof module to be able to print a metric value based on its
type.