Expose HPCRUN_DEFAULT_SAMPLE_RATE for users
The goal is to let users to define the default frequency threshold for perf_event sample source.
For instance if to set 100 as the default frequency, it needs to specify the constant in the CFLAGS:

../configure ... CFLAGS="-DHPCRUN_DEFAULT_SAMPLE_RATE=100"

The sample-rate number idealy has to be less than /proc/sys/kernel/perf_event_max_sample_rate

Note: the name of the constant is inspired by perf_event's "perf_event_max_sample_rate", and it isn't consistent with hpcrun's DEFAULT_THRESHOLD constant defined in some sample-sources files.