refine is_leaf condition to take into account the node type
the datacentric has an attribute node type to distinguish between regular node, leaf node, and new types for datacentric: allocation node (where the malloc statement is located) and memaccess node (where a statement has memory access).
Later on, hpcviewer needs to have a separate view to display the cct of the variables (for dynamic node it's the cct of the malloc), and its memory access. We need to have a separate view to avoid problems with callers view and flat view.