Add publication list with the recent ones
Generating publications.html.in requires bib2xhtml and pubs.bib. The later is generated automatically by bibdesk, but I think it's okay to edit it manually.

Note:
We need to resolve BIB2XHTML_ROOT variable in Makefile to allow users to set easily.
Right now, the value has to be hard-coded.