Address the possiblity that writing a text-formatted "recipe", when it is the libunwind recipe used on an x86, will write the libunwind recipe as if it is an x86 recipe, and print nonsense.
Push an unwinder_t parameter all the way down to the function that
text-formats recipes, and for the x86 version of that function, choose
the formatter type that matches the type of the unwinder.

With regard to conversion-to-string functions, remove some unused ones, make
static some only used near their definitions, inline some trivial ones, and
make conditional the compilation of some that are only invoked by conditionally-
compiled code.