Consistently use void** type for representing and passing x86 register values.
Treat the result of a libunwind step as a failure, even when libunwind thinks
it was a success, when the value of the stack pointer register moves the wrong
way.

After a libunwind step failure on an x86, load the x86 native unwinder
recipe before starting a native unwind step.