Rework the configure options for spack builds for specifying the MPI C++ compiler for hpcprof-mpi.
Replace --enable-cray-mpi with --enable-mpi-search=(cray,bgq) which
tells configure to search for the correct MPI wrapper for this system
type.  This is necessary on Cray and Blue Gene because the normal
spack method for setting mpicxx in packages.yaml doesn't work on these
systems.

On Cray, the MPI wrapper is CC, but spack overrides CC with a C++
compiler, not an MPI compiler.

On Blue Gene, mpicxx is either not available as a module or else not
in a bin directory.

Outside of spack, continue to use MPICXX, plus --enable-all-static if
MPI builds static binaries for the compute nodes.