datacentric: store the list of memory events in a header file
Temporarily store the list of pmu memory-related events in a header file
(pmu_x86.h). This is not very customizable, but it does the purpose.
In the future, we may think to use json (like perf tool) so that we
don't have to tweak the source code every time we want to install to a
new cpu :-(