Rework how the launch scripts find the install directory.
Write the install prefix from configure into the script, and create a
hash value to verify that we've found the directory that matches the
script.  This should robustly cover the two main cases:

  1. As long as the install prefix is the same, then the script can be
  copied anywhere.

  2. As long as the script is in the bin (or libexec) directory, then
  the entire install tree can be moved anywhere.

As a last resort, try the HPCTOOLKIT environ variable.  But that
should only be necessary if both the script and its install directory
have moved.