Merge irreducible loops.
If L1 is an irreducible loop and L2 is an irreducible subloop and L1
and L2 have the same file and line attribution, then merge L2 into L1.
It's likely there is really only one loop in the source code.

This only applies to irreducible loops.  Natural (reducible) loops are
assumed to be correct.