hpcprof: draft to output variable names in experiment.xml
For datacentric event, hpcprof combines the information from hpcstruct
(about static variables) and hpcrun (the vma of the memory access).
Currently, hpcprof treats memory addresses (the variables) the same as
procedures. This works fine if the memory address is recognized either
from static variable or head allocation. If the memory address is from
stack or other unknown source, it will print in xml "unknown procedure",
which is misleading. (it should be "unknown memory" or something like
that).