hpcprof: preparing artificial root for datacentric
To avoid major overhaul and keep backward compatibility, hpcprof needs
to output 2 roots: one for CCT, and the other one for datacentric which
contains the tree for variable (or memory allocation) and its call path
accesses.

Datacentric tree has to be separated from normal CCT bacause (1) its
structure is different; (2) it will be constructed by the viewer in a
different view than top-down view; and (3) we need to allow hpcprof and
hpcprof-mpi to combine with different measurement directories (for
instance, we want to combine datacentric measurement directory with
top-down measurement).

Minor update:
- A static variable must be located in a source file. If Dyninst cannot
expose the source file, then hpcprof will output the source file as
"unknown file".