hpcprof: accumulate inclusive metric up to datacentric root
hpcprof-mpi and hpcprof (if summary metric flag is enabled) will
accumulate inclusive and exclusive metrics up to the "invisible" root.
Since datacentric root is supposed to viewed separated by the viewer, we
don't want to accumulate datacentric metrics to the "invisible" root.
Otherwise we will double count the metrics.

Example xml layout:
<invisible root>  100 %
  <datacentric>    100 %
  <program root>    80 %
  <thread root>     20 %

The layout in the viewer:

<datacentric> 100 %
   <static variable>  40 %
   <heap allocation>  40 %
   <unknown variable> 20 %

<invisible root>  100 %
  <program root>    80 %
  <thread root>     20 %