hpcprof: add <unknown variable> name if vma is not recognized
Instead of displaying "<unknown procedure>" (default), we should disply
"unknown variable" on the viewer if a memory address is not part of
global variable nor heap allocation. It's probably in the stack ?

hpcrun: increment 1 for lm-ip of the static variable. Somehow hpcprof
decrement the value of lm-ip, and thus is not within the known range of
static variables

hpcrun: make "start address" and "end address" metrics hidden.
Unfortunately the naughty hpcprof will not output to xml all hidden
metrics.
The start and end memory address metrics are also important for
memory-centric (to be integrated with data-centric).
I think hpcprof shouldn't remove hidden metrics.