hpcprof: mark the datacentric root for the viewer
If a datacentric is enabled, the experiment.xml will contain two trees:
one is CCT (the traditional one), the other one is datacentric tree.
To inform the viewer a datacentric tree has to be rendered separatedly,
we can either output it in a different xml file (preferable), or embed
write a marker.

The former requires major surgery to hpcprof, while the later only
requires minor changes in hpcprof.

Minor fix:
- hpcprof has a major issue with hpcrun's hidden metrics. It will remove
all the first hidden metrics until it reaches visible ones (since it has
a vector of temporary metrics that are hidden). To avoid this, we should
mark all hpcrun metrics as "visible", or otherwise they are missing in
xml file.