hpcrun: draft including address centric in data-centric event
To record the first touch, we use perf_events' PAGE-FAULTS event which
is delivered when a page fault occurs (by the kernel).

It seems this software event has the same effect with mprotect()
but without requiring to handle sigsegv. We need to avoid handling
sigsegv since it interferes with hpcrun sigsegv. It's just too
problematic to use.