fix failure to notify unwind intervals when remapping same file
synprog is a test program that calls functions in shared libraries.
it maps a shared library, uses it, unmaps it, remaps it, and uses
it again. when the library was remapped, the load map was properly
updated. however, the loadmap failed to issue "map" notification
callbacks when the named shared library had already been in the
loadmap. this commit ensures that the "map" notification occurs
even if the shared library name is known to the loadmap.