Datacentric: temporary fix for page faults
Temporarily, not to sample every page-faults, but instead sample for a
certain frequency. We should use mprotect instead of perf_events'
page-fault to enable mask and unmask the memory protection. However,
this requires careful handle of sigsegv. (and we want to avoid tightly
couple datacentric and hpcrun since datacentric is a "plugin".


Minor fix:
- hpcprof: outpur the period mean as float
- fix hpcfnbounds to only send static variables that match datacentric
minimum bytes requirement. (sharing code between hpcrun-datacentric and
hpcfnbounds in support-lean).
- perf-util: make perf_event to wake up for every overflows.