add handling for a corner case to the ompt frame elider
add handling for corner case where a thread is not in the idle
state, but the whole call stack is elided anyway. this can
occur if sampled between the time a thread's state is adjusted
to indicate the thread is beginning parallel work but
before the a thread begins the parallel work.

rather than returning an empty call stack, collapse the call
stack to <openmp idle>.