Disable first-touch/address centric feature temporarily
The plan for data centric is to support address centric (to get the memory location of the access) and to campture first-touch policy.
Howver, the later needs to use mprotect which requires to handle sigsegv. This sigsegv is already handled by hpcrun, and datacentric needs to be very careful if it wants to recover the signal.
On the other hand, using perf_events' page-faults causes to signal all page faults, whether we are interested of the memory reference or not.

In this commit, datacentric only captures memory latency, page faults and the memory size of some variables.