hpcrun: Simplify kernel block time event
Instead of tweaking the code to handle Linux kernel bug for monitoring
context switch event for sleep or other clock-based routines, we just
assume the block time of those routines is negligeable.

see https://www.mail-archive.com/linux-kernel@vger.kernel.org/msg1424933.html

(this bug only occurs for self profilers like hpcrun, not 3rd party
profiler like Linux perf tool)

A simpler solution is to focus on context-switch out and in, and to
ensure the block time is accurate for all kernel blocks time, except
inside clock-related routines.

The accuracy of the original code depends on the high frequency of other
events (like cycles). If this frequency is low, the block time is
inaccurate.

This commit fixes this problem, but it sacrifies the block time inside
sleep or other clock-related routines.

Let's hope Linux kernel developer fixes this bug.