turn hpcio_outbuf_t into an opaque type
with the fields of hpcio_outbuf_t defined in a header file
and including a spinlock defined with our <stdatomic.h> header,
this caused a conflict with boost atomics used by dyninst.
the best way to avoid this conflict was to make hpcio_outbuf_t
and opaque type so the include of <spinlock.h> and <stdatomic.h>
could be in hpcio-buffer.c rather than hpcio-buffer.h.

with hpcio_outbuf_t as an opaque type, it couldn't be a member
of core_profile_trace_data.h. instead, the member became a pointer
to hpcio_outbuf_t. that required methods to allocate and free
hpcio_outbuf_t that could be invoked concurrently.

a pile of code to fix a macro conflict :-(