hpcprof-mpi: Fix issue #139
The problem is that CoefVarIncr class doesn't overide method numAccum() needed to return the number of accumulators needed for the computation (by default it's set 1 by the parent).
Hence, when parallel reduce asks the number of accumulators, the answer is always 1.

Overriding numAccumi() fixes this issue.
Also removing hasAccum() in the same class since nobody uses it.