hpcprof: Fix bug for duplicate inline procedures
We use triplet <load_module, file, procedure_name> to ensure a procedure is unique.
However, since we we use the "ancestor file" of a procedure, for an inline procedure, it points to the file where it is used, not where it is defined.
By using file where an inline is defined, this fixes duplicity for inlines.