hpcrun datacentric: link cct node to variable node
Add information in hpcrun to help hpcviewer to link between a node in memory access to a node in datacentric tree (the node where the variable is declarated or allocated).

In datacentric tree, we can easily navigate from an allocation node, to its memory accesses.
However, we currently don't know how to navigate between top-down or bottom-up or flat tree to datacentric tree. T
To help hpcviewer, hpcrun needs to add more information to navigate between cct and datacentric tree.