perf_events: log throtted and lost samples
If we set too many events (too many multiplexed events) with too high frequency, the kernel may need to throttle samples, which may cause inaccuracies (and the result doesn't make sense).
We need to log how many throttles and lost samples so we can set or adjust the frequency.

It will be better if hpcrun can adjust dynamically the frequency.