perf_events: filtering precise ip before adding it in the cct
Sometimes PEBS precise ip is completely different than context ip. For
example:

PEBS precise ip: 0xffffffff9421181a (lm: -1)
signal's context ip: 0x400971 (lm: 1)

They can be so different, that even the load module is not the same.
In this case, we add the precise ip node as the child node.